producer/requirements.txt
pandas
kafka-python
tqdm
